create player module
