+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:name="com.jeanpaulo.musiclibrary.settings.ui.SettingsActivity"
             android:exported="false" />
 
-        <service android:name="com.jeanpaulo.musiclibrary.core.music_player.MPService" />
+        <service android:name="com.jeanpaulo.musiclibrary.player.mp.MPService" />
 
     </application>
