--- core/src/main/java/com/jeanpaulo/musiclibrary/core/music_player/MPParams.kt
@@ -1,10 +1,0 @@
-package com.jeanpaulo.musiclibrary.core.music_player
-
-object MPParams {
-    const val PLAYLIST = "playlist"
-    const val COMMAND = "command"
-    const val SONG = "song"
-
-    const val COUNTER = "counter"
-    const val STATE = "state"
-}
