--- core/src/main/java/com/jeanpaulo/musiclibrary/core/presentation/SimpleMusicDetailUIModel.kt
@@ -1,11 +1,0 @@
-package com.jeanpaulo.musiclibrary.core.presentation
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class SimpleMusicDetailUIModel(
-    val id: Long,
-    val name: String,
-    val artworkUrl: String
-): Parcelable
