--- music/ui/src/main/res/menu/menu_music_detail.xml
@@ -1,14 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".app.music_detail.presentation.MusicDetailActivity">
-    <item
-        android:id="@+id/action_favorite"
-        android:checkable="true"
-        android:icon="@drawable/check_favorite"
-        android:orderInCategory="100"
-        android:title="@string/action_favorite"
-        android:visible="true"
-        app:showAsAction="ifRoom" />
-    
-</menu>
