+++ player/src/main/java/com/jeanpaulo/musiclibrary/player/mp/MPCommands.kt
@@ -1,0 +1,17 @@
+package com.jeanpaulo.musiclibrary.player.mp
+
+object MPCommands {
+    const val PLAY = "play"
+    const val PLAY_SONG = "play_song"
+    const val PLAY_SONG_LIST = "play_song_list"
+
+    const val PAUSE = "pause"
+    const val STOP = "stop"
+
+    const val NEXT = "next"
+    const val PREVIOUS = "previous"
+
+    const val UPDATE_COUNTER = "update_counter"
+
+    const val ADD_SONG = "add_song"
+}
