+++ player/src/main/java/com/jeanpaulo/musiclibrary/player/mp/MPParams.kt
@@ -1,0 +1,10 @@
+package com.jeanpaulo.musiclibrary.player.mp
+
+object MPParams {
+    const val PLAYLIST = "playlist"
+    const val COMMAND = "command"
+    const val SONG = "song"
+
+    const val COUNTER = "counter"
+    const val STATE = "state"
+}
