+++ player/src/main/res/layout/full_player_bottom_sheet.xml
