+++ player/src/main/res/layout/mini_player_bottom_sheet.xml
