+++ player/src/main/res/values/strings.xml
